refactor(cli): remove exports from snapshot files
