+++ core/monorule/src/cli/index.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { run } from "./run";
+
+await run();
