+++ core/monorule/tsconfig.json
@@ -8,7 +8,8 @@
   ],
   "compilerOptions": {
     "paths": {
-      "@/*": ["./*"]
+      "~/*": ["./*"],
+      "@/*": ["./src/*"]
     }
   }
 }
